+++ .github/workflows/playerData.yml
@@ -1,0 +1,31 @@
+name: Update Player Data
+
+on:
+  schedule:
+    - cron: '0 0 * * *'  # 毎日午前0時に実行
+  workflow_dispatch:  # 手動実行のオプション
+
+jobs:
+  update-player-data:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Set up Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: '20'
+
+      - name: Install dependencies
+        run: npm ci
+
+      - name: Run getPlayerData script
+        run: node scripts/getPlayerData.mjs --only-player-json
+
+      - name: Commit and push if changed
+        run: |
+          git config --local user.email "action@github.com"
+          git config --local user.name "GitHub Action"
+          git add data/player.json
+          git diff --quiet && git diff --staged --quiet || (git commit -m "Update player data" && git push origin dev)
