+++ data/player.json
@@ -1,0 +1,1 @@
+{"playerId":"bbc52cb2","signatureTimestamp":"19997"}
