--- examples/Serverless/Vercel/.vercel/cache/node/api/index.ts/tsconfig-with-tsconfig-json.json
@@ -1,6 +1,0 @@
-{
-	"extends": "..\\..\\..\\..\\..\\tsconfig.json",
-	"include": [
-		"..\\..\\..\\..\\..\\api\\index.ts"
-	]
-}
