--- examples/Serverless/Vercel/.vercel/cache/node/api/index.ts/tsconfig.json
@@ -1,5 +1,0 @@
-{
-	"include": [
-		"..\\..\\..\\..\\..\\api\\index.ts"
-	]
-}
