--- examples/Serverless/Vercel/.vercel/project.json
@@ -1,1 +1,0 @@
-{"projectId":"prj_IaFrw1DG6jz2Clj8TydfADYdo6fz","orgId":"zkiTMNHEPr6kemeQj4it2tKb"}
