+++ package.json
@@ -58,7 +58,7 @@
     "scripts": {
         "test": "npx jest ./test/main.test.ts --detectOpenHandles",
         "clear-cache-files": "cd package/core && rmdir /s /q CacheFiles 2>nul & cd ../..",
-        "build": "node ./scripts/getPlayerId.mjs && rmdir /s /q package & tsc && tsc-alias && npm run clear-cache-files && npm run create-bundles",
+        "build": "node ./scripts/getPlayerData.mjs && rmdir /s /q package & tsc && tsc-alias && npm run clear-cache-files && npm run create-bundles",
         "publish:npm": "node ./scripts/publish.mjs latest && npm run build && npm publish --registry=https://registry.npmjs.org/",
         "publish:npm-alpha": "node ./scripts/publish.mjs alpha && npm run build && npm publish --registry=https://registry.npmjs.org/ --tag alpha",
         "publish:npm-beta": "node ./scripts/publish.mjs beta && npm run build && npm publish --registry=https://registry.npmjs.org/ --tag beta",
