+++ src/utils/Constants.ts
@@ -1,6 +1,6 @@
 import pkg from '../../package.json';
 
-export const CURRENT_PLAYER_ID = 'd9418494';
+export const CURRENT_PLAYER_ID = 'bbc52cb2';
 export const VERSION = pkg.version;
 export const REPO_URL = 'https://github.com/ybd-project/ytdl-core';
 export const ISSUES_URL = 'https://github.com/ybd-project/ytdl-core/issues';
